+++ src/functions/fmt.ts
@@ -4,14 +4,14 @@
 
 export async function fmtYapf(
   pyTxt: string,
-  { innerText, based_on_style = 'pep8', column_limit = 120 },
+  { based_on_style = 'pep8', column_limit = 120 },
   meta = {}
-): Promise<string> {
+): Promise<string | undefined> {
   /**
    * Format Python code with YAPF. Of course, YAPF needs to be installed.
    * YAPF is called within a Shell to allow it to read local config files, ENV options, etc.
    */
-  let text = innerText.trim() || pyTxt.trim();
+  let text = pyTxt.trim();
   if (!text) return;
   const opts = `{based_on_style:${based_on_style}, column_limit:${column_limit}}`;
   // TODO: call user's shell instead of ZSH
@@ -21,7 +21,7 @@ export async function fmtYapf(
   return text;
 }
 
-export async function fmtBlack(pyTxt: string, args, meta = {}): Promise<string> {
+export async function fmtBlack(pyTxt: string, args: any, meta = {}): Promise<string | undefined> {
   /**
    * Format Python code with Black. Of course, Black needs to be installed.
    * Black is called within a Shell to allow it to read local config files, ENV options, etc.
@@ -29,7 +29,7 @@ export async function fmtBlack(pyTxt: string, args, meta = {}): Promise<string> 
   return await fmtBlackOrBlue('black', pyTxt, args, meta);
 }
 
-export async function fmtBlue(pyTxt: string, args, meta = {}): Promise<string> {
+export async function fmtBlue(pyTxt: string, args: any, meta = {}): Promise<string | undefined> {
   /**
    * Format Python code with Blue. Of course, Blue needs to be installed.
    * Blue is called within a Shell to allow it to read local config files, ENV options, etc.
@@ -39,11 +39,11 @@ export async function fmtBlue(pyTxt: string, args, meta = {}): Promise<string> {
 
 async function fmtBlackOrBlue(
   exec: string,
-  pyTxt: string,
-  { innerText, line_length = 100 },
+  text: string,
+  { line_length = 100 },
   meta = {}
-): Promise<string> {
-  let text = innerText.trim() || pyTxt.trim();
+): Promise<string | undefined> {
+  text = text.trim();
   if (!text) return;
   // TODO: call user's shell instead of ZSH
   const proc = Bun.spawn(['zsh', '-c', `${exec} --quiet --stdin-filename script.py --line-length=${line_length} -`], {
@@ -56,12 +56,12 @@ async function fmtBlackOrBlue(
   return text;
 }
 
-export async function fmtPrettier(jsTxt: string, { innerText, print_width = 120 }, meta = {}): Promise<string> {
+export async function fmtPrettier(text: string, { print_width = 120 }, meta = {}): Promise<string | undefined> {
   /**
    * Format Javascript code with Prettier. Of course, Prettier needs to be installed.
    * Prettier is called within a Shell to allow it to read local config files.
    */
-  let text = innerText.trim() || jsTxt.trim();
+  text = text.trim();
   if (!text) return;
   // TODO: user's shell instead of ZSH
   const proc = Bun.spawn(['zsh', '-c', `bunx prettier --print-width ${print_width} --stdin-filepath script.js`], {