+++ src/functions/string.ts
@@ -1,32 +1,32 @@
-export function lower(text: string, { innerText }): string {
+export function lower(text: string): string {
   /**
    * Lower-case all the text.
    */
-  return (innerText || text).toLowerCase();
+  return text.toLowerCase();
 }
 
-export function upper(text: string, { innerText }): string {
+export function upper(text: string): string {
   /**
    * Upper-case all the text.
    */
-  return (innerText || text).toUpperCase();
+  return text.toUpperCase();
 }
 
-export function titleAll(text: string, { innerText }): string {
+export function titleAll(text: string): string {
   /**
    * Title case for all the words.
    * It would be nice if this was called just "title", but
    * there is an HTML tag called "title" already.
    */
   // text.toLowerCase().replace(/(?=\b)(\w)/g, (m, $1) => $1.toUpperCase());
-  return (innerText || text).replace(/\w\S*/g, txt => txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase());
+  return text.replace(/\w\S*/g, txt => txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase());
 }
 
-export function trim(text: string, { innerText }): string {
+export function trim(text: string): string {
   /**
    * Trim whitespaces from both ends of the text.
    */
-  return (innerText || text).trim();
+  return text.trim();
 }
 
 export function line(len: string, { c = '-' } = {}): string {
@@ -39,12 +39,11 @@ export function line(len: string, { c = '-' } = {}): string {
   return c.repeat(nr);
 }
 
-export function sortLines(text: string, { innerText, caseSensitive = false } = {}): string {
+export function sortLines(text: string, { caseSensitive = false } = {}): string | undefined {
   /**
    * Sort lines of text alphabetically.
    * By default, the sorting is case insensitive.
    */
-  text = innerText || text;
   if (!text) return;
 
   let sortFunc = null;