+++ src/functions/xeval.ts
@@ -5,12 +5,12 @@ import vm from 'vm';
 
 import { splitToMarker } from '../util.ts';
 
-export function jsEval(zeroExpr: string, args: Record<string, string> = {}) {
+export function jsEval(expr: string, args: Record<string, string> = {}) {
   /**
    * Eval JavaScript and return the result.
    * This uses the builtin eval function from Bun.
    */
-  let expr = (zeroExpr || args.innerText || args.expr).trim();
+  expr = (expr || args.expr).trim();
   if (!expr) return;
   expr = splitToMarker(expr);
   const stdout: string[] = [];
@@ -33,12 +33,12 @@ ${stdout.length ? stdout.join('\n') : ''}${result === undefined ? '' : '\n' + re
 `;
 }
 
-export async function pyEval(zeroExpr: string, args: Record<string, any> = {}) {
+export async function pyEval(expr: string, args: Record<string, any> = {}) {
   /**
    * Eval Python and return the result. Useful for Math.
    * Python is installed in most Linux distributions and on MacOS.
    */
-  let expr = (zeroExpr || args.innerText || args.expr).trim();
+  expr = (expr || args.expr).trim();
   if (!expr) return;
   expr = splitToMarker(expr);
   const proc = Bun.spawn(['python3'], { stdin: 'pipe', stdout: 'pipe', stderr: 'pipe' });
@@ -62,24 +62,24 @@ ${stdout.trim()}${stderr.trim()}
 `;
 }
 
-export async function rbEval(zeroExpr: string, args: Record<string, any> = {}) {
+export async function rbEval(expr: string, args: Record<string, any> = {}) {
   /**
    * Eval Ruby expression and return the result. Useful for Math.
    */
-  const expr = zeroExpr || args.innerText || args.expr;
-  if (!expr || !expr.trim()) return;
+  expr = (expr || args.expr).trim();
+  if (!expr) return;
   const print = args.print === undefined || args.print === true ? `p ${expr}` : expr.toString();
   const proc = Bun.spawn(['ruby', '-e', print]);
   const result = await new Response(proc.stdout).text();
   return result.trim();
 }
 
-export async function perlEval(zeroExpr: string, args: Record<string, any> = {}) {
+export async function perlEval(expr: string, args: Record<string, any> = {}) {
   /**
    * Eval Perl expression and return the result. Useful for Math.
    */
-  const expr = zeroExpr || args.innerText || args.expr;
-  if (!expr || !expr.trim()) return;
+  expr = (expr || args.expr).trim();
+  if (!expr) return;
   const say = args.say === undefined || args.say === true ? `say ${expr}` : expr.toString();
   const proc = Bun.spawn(['perl', '-E', say]);
   const result = await new Response(proc.stdout).text();